Updated so file writes regularly instead of at end
Since the write function occurs after running through the entire list, if problems are encountered during the run then all progress is lost. This isn't a big deal on short runs, but on large lists, it can be problematic. I have moved the write function so that VALIDs are written immediately upon detection.